+++ sklearn/preprocessing/__init__.py
@@ -28,7 +28,7 @@ from .label import LabelBinarizer
 from .label import LabelEncoder
 from .label import MultiLabelBinarizer
 
-from .target import TransformedTargetRegressor
+from .target import TransformTargetRegressor
 
 from .imputation import Imputer
 
@@ -47,7 +47,7 @@ __all__ = [
     'OneHotEncoder',
     'RobustScaler',
     'StandardScaler',
-    'TransformedTargetRegressor',
+    'TransformTargetRegressor',
     'add_dummy_feature',
     'PolynomialFeatures',
     'binarize',